+++ test/distance/tests-OSA.cpp
@@ -7,20 +7,20 @@
 #include "../common.hpp"
 
 template <typename Sentence1, typename Sentence2>
-int64_t osa_distance(const Sentence1& s1, const Sentence2& s2,
-                     int64_t max = std::numeric_limits<int64_t>::max())
+size_t osa_distance(const Sentence1& s1, const Sentence2& s2,
+                     size_t max = std::numeric_limits<size_t>::max())
 {
-    int64_t res1 = rapidfuzz::osa_distance(s1, s2, max);
-    int64_t res2 = rapidfuzz::osa_distance(s1.begin(), s1.end(), s2.begin(), s2.end(), max);
-    int64_t res3 = rapidfuzz::osa_distance(
+    size_t res1 = rapidfuzz::osa_distance(s1, s2, max);
+    size_t res2 = rapidfuzz::osa_distance(s1.begin(), s1.end(), s2.begin(), s2.end(), max);
+    size_t res3 = rapidfuzz::osa_distance(
         BidirectionalIterWrapper(s1.begin()), BidirectionalIterWrapper(s1.end()),
         BidirectionalIterWrapper(s2.begin()), BidirectionalIterWrapper(s2.end()), max);
     rapidfuzz::CachedOSA scorer(s1);
-    int64_t res4 = scorer.distance(s2, max);
-    int64_t res5 = scorer.distance(s2.begin(), s2.end(), max);
+    size_t res4 = scorer.distance(s2, max);
+    size_t res5 = scorer.distance(s2.begin(), s2.end(), max);
 #ifdef RAPIDFUZZ_SIMD
     if (s1.size() <= 64) {
-        std::vector<int64_t> results(256 / 8);
+        std::vector<size_t> results(256 / 8);
 
         if (s1.size() <= 8) {
             rapidfuzz::experimental::MultiOSA<8> simd_scorer(1);
